fixed the reason why build failed (removed unused bean)

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/librarysystem/service/HeadLibrarianService.java
@@ -4,13 +4,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import ca.mcgill.ecse321.librarysystem.dao.*;
+import ca.mcgill.ecse321.librarysystem.dao.AccountRepository;
+import ca.mcgill.ecse321.librarysystem.dao.EventRepository;
+import ca.mcgill.ecse321.librarysystem.dao.MediaRepository;
+import ca.mcgill.ecse321.librarysystem.dao.OpeningHourRepository;
+import ca.mcgill.ecse321.librarysystem.dao.ShiftRepository;
 import ca.mcgill.ecse321.librarysystem.models.Account;
 import ca.mcgill.ecse321.librarysystem.models.Account.AccountCategory;
 import ca.mcgill.ecse321.librarysystem.models.HeadLibrarian;
 import ca.mcgill.ecse321.librarysystem.models.Librarian;
 import ca.mcgill.ecse321.librarysystem.models.Shift;
-import javassist.expr.Instanceof;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -29,10 +32,21 @@ public class HeadLibrarianService {
     @Autowired
     ShiftRepository shiftRepository;
     
+    /**
+     * Create a head librarian with given parameters
+     * @param aId
+     * @param aAddress
+     * @param aName
+     * @param accountCategory
+     * @param local
+     * @param itemsChecked
+     * @return the head librarian created
+     */
+    
     @Transactional
     public HeadLibrarian createHeadLibrarian(int aId, String aAddress, String aName, 
     		AccountCategory accountCategory, boolean local, int itemsChecked) {
-    	/*if (aId==0) {
+    	if (aId==0) {
             throw new IllegalArgumentException("Head Librarian id cannot be 0.");
         }
     	if (aAddress==null) {
@@ -46,7 +60,7 @@ public class HeadLibrarianService {
     	}
     	if (local==false) {
     		throw new IllegalArgumentException("Head Librarian must be a local");
-    	}*/
+    	}
     	HeadLibrarian head=new HeadLibrarian();
     	head.setId(aId);
     	head.setAddress(aAddress);
@@ -56,29 +70,60 @@ public class HeadLibrarianService {
     	head.setNumChecked(itemsChecked);
 		return head;
     }
-    
+    /**
+     * Find head librarian with given parameter
+     * @param aId
+     * @return the head librarian
+     */
     @Transactional
     public HeadLibrarian getHeadLibrarian(int aId) {
     	return (HeadLibrarian) accountRepository.findAccountById(aId);
     }
     
+    /**
+     * Find a head librarian and change their info with given parameters
+     * @param aId
+     * @param aAddress
+     * @param aName
+     * @param itemsChecked
+     */
     @Transactional
-    public void updateHeadLibrarianInfo(int aId, String aAddress, String aName, int itemsChecked) {
+    public HeadLibrarian updateHeadLibrarianInfo(int aId, String aAddress, String aName, int itemsChecked) {
     	HeadLibrarian head=(HeadLibrarian) accountRepository.findAccountById(aId);
     	head.setAddress(aAddress);
     	head.setName(aName);
+		return head;
     }
     
+    /**
+     * Delete head librarian corresponding to the given parameter
+     * @param aId
+     */
     @Transactional
-    public void deleteHeadLibrarian(int aId) {
-    	accountRepository.delete(accountRepository.findAccountById(aId));
+    public HeadLibrarian deleteHeadLibrarian(int aId) {
+    	HeadLibrarian head=(HeadLibrarian) accountRepository.findAccountById(aId);
+    	shiftRepository.deleteAll(shiftRepository.findByLibrarian((Librarian) head));
+    	mediaRepository.deleteAll(mediaRepository.findByAccount(head));
+    	eventRepository.deleteAll(eventRepository.findByAccount(head));
+    	accountRepository.delete(head);
+		return head;
     }
     
+    /**
+     * Create librarian with given parameters
+     * @param aId
+     * @param aAddress
+     * @param aName
+     * @param accountCategory
+     * @param local
+     * @param itemsChecked
+     * @return
+     */
     @Transactional
     public Librarian hireLibrarian(int aId, String aAddress, String aName, 
     		AccountCategory accountCategory, boolean local, int itemsChecked) {
     	//i.e. createLibrarian
-    	/*if (aId==0) {
+    	if (aId==0) {
     		throw new IllegalArgumentException("Librarian id cannot be 0.");
     	}
     	if (aAddress==null) {
@@ -92,7 +137,7 @@ public class HeadLibrarianService {
     	}
     	if (local==false) {
     		throw new IllegalArgumentException("Librarian must be a local");
-    	}*/
+    	}
     	Librarian librarian=new Librarian();
     	librarian.setId(aId);
     	librarian.setAddress(aAddress);
@@ -103,18 +148,27 @@ public class HeadLibrarianService {
 		return librarian;
     }
     
+    /**
+     * Delete librarian corresponding to given parameter
+     * @param aId
+     */
 	@Transactional
-    public void fireLibrarian(int aId) {	//i.e. deleteLibrarian
-    	/*if (aId==0) {
-    		throw new IllegalArgumentException("There is not a id 0 account.");
-    	}*/
+    public Librarian fireLibrarian(int aId) {	//i.e. deleteLibrarian
+    	if (aId==0) {
+    		throw new IllegalArgumentException("There is not an account with id 0.");
+    	}
 		Account firedLibr=accountRepository.findAccountById(aId);
     	shiftRepository.deleteAll(shiftRepository.findByLibrarian((Librarian) firedLibr));
     	mediaRepository.deleteAll(mediaRepository.findByAccount(firedLibr));
     	eventRepository.deleteAll(eventRepository.findByAccount(firedLibr));
     	accountRepository.delete(firedLibr);
+		return (Librarian) firedLibr;
     }
 	
+	/**
+	 * Find all librarians
+	 * @return list of all librarians
+	 */
 	@Transactional
 	public List<Librarian> getLibrarians(){
 		List<Account> allAccounts=(List<Account>) accountRepository.findAll();
